+++ src/ts/ir/index.ts
@@ -149,9 +149,9 @@ class IR {
             // 打开链接
             const aElement = hasClosestByAttribute(event.target, "data-type", "a");
             if (aElement && (!aElement.classList.contains("vditor-ir__node--expand"))) {
-                if (vditor.options.link.trigger) {
-                    vditor.options.link.trigger(aElement.querySelector(":scope > .vditor-ir__marker--link"));
-                } else if (vditor.options.link.open) {
+                if (vditor.options.link.click) {
+                    vditor.options.link.click(aElement.querySelector(":scope > .vditor-ir__marker--link"));
+                } else if (vditor.options.link.isOpen) {
                     window.open(aElement.querySelector(":scope > .vditor-ir__marker--link").textContent);
                 }
                 return;